% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_proxy.R
\name{init_proxy}
\alias{init_proxy}
\title{Initialization proxy settings}
\usage{
init_proxy()
}
\value{
a file in home directory
}
\description{
When user use it in the first time or reset proxy, this will run and store setting in home directory.
}
\examples{
init_proxy()
}
